requirements.txt
tensorflow>=2.10
numpy
pillow
matplotlib
scikit-learn
streamlit
opencv-python
